w3c-reports/12092014/mac/runned-tests/dom/nodes/Document-createTreeWalker-expected.txt


FAIL Required arguments to createTreeWalker should be required. assert_throws: function "function () {
    document.createTreeWalker();
  }" threw object "Error: NotSupportedError: DOM Exception 9" ("NotSupportedError") expected object "TypeError" ("TypeError")(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:593:23
http://localhost:8800/dom/nodes/Document-createTreeWalker.html:9:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/dom/nodes/Document-createTreeWalker.html:8:5)
FAIL Optional arguments to createTreeWalker should be optional (1 passed). assert_equals: expected 4294967295 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-createTreeWalker.html:17:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/dom/nodes/Document-createTreeWalker.html:13:5)
FAIL Optional arguments to createTreeWalker should be optional (2 passed). assert_equals: expected null but got object "[object NodeFilter]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-createTreeWalker.html:25:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/dom/nodes/Document-createTreeWalker.html:20:5)
FAIL Optional arguments to createTreeWalker should be optional (3 passed, null). assert_equals: expected null but got object "[object NodeFilter]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-createTreeWalker.html:32:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/dom/nodes/Document-createTreeWalker.html:27:5)
FAIL Optional arguments to createTreeWalker should be optional (3 passed, function). assert_equals: expected (function) function "function () {}" but got (object) object "[object NodeFilter]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/dom/nodes/Document-createTreeWalker.html:40:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/dom/nodes/Document-createTreeWalker.html:34:5)
